net: eqos: check for context descriptors when calculating tx_bytes
Prior to this change, the tx completion code blindly added
the descriptor length to the total transmitted bytes (tx_bytes)
count. This is broken when context descriptors are used (e.g. VLAN
tagging is enabled) because they set their length to -1. Since that
field is an unsigned short, the operation resulted in 64K bytes added
to tx_bytes. With high data rates, this is quickly noticed as > 1Gbps
of data being transmitted on the underlying interface.

Bug: 36090351
Bug 1893981

Change-Id: I6d93403cbe54901121649c53f5a175349bd7cab0
Signed-off-by: Dima Zavin <dmitriyz@google.com>
Reviewed-on: http://git-master/r/1331239
Reviewed-on: https://git-master.nvidia.com/r/1517715
GVS: Gerrit_Virtual_Submit
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Bhadram Varka <vbhadram@nvidia.com>
Tested-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>