DT: t18x: common: Use denver DT odm data for WDT enable
There is two watchdog timer called AP and denver. The AP watchdog
timer is not used and it is recommended to use the denver WDT
only. With this recommendation, AP WDT is removed and only Denver
WDT is used for watchdog purpose.

Enable the Denver WDT with the odm data of "enable-denver-wdt".

Later flag for tegra-wdt will be deprecated.

Change-Id: I8e741bb3c5f38b396cc20baffc9ade0644ecc956
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1454065
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>