ethernet: eqos: register fixed PHY
Few boards are using the EQOS driver does not have a conventional
PHY attached to it, it is directly connected to a somewhat simple
ethernet switch. This switch does not present the usual register
interface of a PHY, it presents nothing. So a PHY scan will find
nothing - it finds ID's of 0 for each PHY on the attached MII bus.

Add code support to use a fixed phy if no PHY is found on the MII bus.

Bug 1894484
Bug 1889409
Bug 200302592

(cherry-picked from commit d52e78cacb707ffefe1fc213e1c3c939ecff1813)

Change-Id: I822f44ecddafe480e79059091784a82cc35212eb
Signed-off-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-on: http://git-master/r/1327711
Signed-off-by: dmitry pervushin <dpervushin@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1517280
GVS: Gerrit_Virtual_Submit
Reviewed-by: Amlan Kundu <akundu@nvidia.com>
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>