dts: t19x: bpmp: change clk provider
Now that we have a number of clks supported by bpmp firmware, start
using the "staged" clock provider. The staged clk provider will first
check with bpmp to see if the asking clk is supported and use it if it
is supported. Otherwise, it will create and use a dummy clk (with nops
clk ops).

JIRA BPMP-429

Change-Id: I23a0e3c3bcede2bed9a817a7288d05da6a415401
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/1456392
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>