dc: use tegra_dc_is_powered() over direct pg API
tegra_dc_is_powered() is a utility wrapper that calls
into the appropriate internal function depending on
which SOC we're running on (compile-time). The driver
should be designed in such a way that it always goes
through this interface instead of directly calling
tegra_powergate_is_powered(), especially for logic
that's common to all SOCs.

TDS-1545

Signed-off-by: Shu Zhong <shuz@nvidia.com>

Change-Id: Ib294a4be09805253683de4b2976810cb4758bdbd
Reviewed-on: http://git-master/r/1460989
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Tested-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>