tegra-se:remove unused chained parameter
chained parameter was added as part of the
fix for Bug 1010604 but not used as such.
However, the motive to fix the above bug is
still part of the code.
Just removing the parameter which is not
necessary for this fix

Bug 1010604

Change-Id: Ic505b69d804998bc1285a453ce7ec3f80090b1da
(cherry picked from commit e068e024a9ec6e1ae7a146d808bb84004abb71ef)
Signed-off-by: Shravani Dingari <shravanid@nvidia.com>
Reviewed-on: http://git-master/r/1461102
Reviewed-on: https://git-master.nvidia.com/r/1530582
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mallikarjun Kasoju <mkasoju@nvidia.com>
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>