pwm: tegra: Read PWM clock source rate in driver init
It is required to know the PWM clock source frequency to
calculate the PWM period.

In driver, the clock source frequency of the PWM does not
get change and, hence, get the clock source frequency in
driver initi. Get this values later for period calculation
from pwm_config().

This will help in avoiding the clock call for getting clock rate
in the pwm_config() each time.

Change-Id: Ied0bd5aedfc0bccab05f9222c183fae03eacb6fd
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>
Reviewed-on: http://git-master/r/1463864
GVS: Gerrit_Virtual_Submit