dc: add hooks for clk switch setup
On T19x, additional steps need to be taken when
switching between the SOR safe vs. macro feedback
clocks. This change adds an inline wrapper function
to account for this, and updates the HDMI and DP
drivers to call this wrapper in the appropriate places.

TDS-2129

Change-Id: Ie07deb721df927db31a32fa73f8fa2f1c738dd7d
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1463131
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aron Wong <awong@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>