video: tegra: fb: remove useless iounmap()
Commit 6668683f (video: tegra: dc: manage alloc/release
framebuffer memory by driver) replaced ioremap() with
other functions while it forgot to remove the iounmap()
in the exit path. When the driver fails during probe(),
it will trigger kernel warnings:
 trying to free invalid coherent area: ffffff80028c1000
 ------------[ cut here ]------------
 WARNING: at arch/arm64/mm/dma-mapping.c:2808
 [...]
 Call trace:
 [<ffffffc00009851c>] arm_iommu_free_attrs+0x58/0x2d8
 [<ffffffc000549cec>] tegra_fb_release_fbmem+0xa8/0xd8
 [<ffffffc00054a934>] tegra_fb_register+0x6dc/0x738
 [<ffffffc0004ea0ac>] tegra_dc_probe+0x9b8/0xf70

This is because the memory space has been through the
iounmap() routine which shouldn't have happened.

So this patch just removes the useless iounmap to fix
the problem.

Bug 1887171
Bug 200289389

Change-Id: I95feff876769970a4bab993d20952c08526bdba7
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1463242
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>