t194: display: update head clock listing
emc_la_clk is not programmed directly from display driver for
NvDisplay. Remove emc_la_clk listing. Add plld4 to the listing
and rearrange the order in which clocks are listed.

Note: TEGRA194_CLK_PLLD3 is causing crash in VDK when tests are run
back-to-back. Replace TEGRA194_CLK_PLLD3 with TEGRA194_CLK_PLLD2
until issue is fixed so that test coverage is not impacted.

tds-502

Change-Id: If73871f9aa677de7c7db033a6e7e429cadbc035f
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: http://git-master/r/1463362
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>