media: i2c: ov9281: Allow initial frame length
Modify the workaround for the instability seen when changing the frame
length during streaming.

In fsync mode, allow the initial frame length setting, which is important
for allowing upper layers to determine the correct frame length for a
particular frame rate.

Previously, we blocked all frame length settings, including the initial
one, when fsync (master or slave) mode is enabled.

AV-348

Change-Id: I847b443ff67522b33f2f85d7430bc352f3f29fc4
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: http://git-master/r/1467122
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>