tegra: dc: hda: fix crash due to invalid dereference
hda context for multiple sor's is allocated in hda_inst and each
instance is referred using index. When hda_inst is freed and context
is referred for sor's other than 0, it returns the offset of that
instance which is not null whereas we are checking for instance
pointer and dereferencing its member which results in crash.

Bug 200301085

Change-Id: I01c95b87e71ab200b10b9b3035190f97d0c49a62
Signed-off-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-on: http://git-master/r/1467522
Reviewed-by: Santosh Galma <galmar@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sumit Bhattacharya <sumitb@nvidia.com>