media: pca9570: remove useless imx185_of_match
If selecting CONFIG_I2C_IOEXPANDER_PCA9570=m,
there is a compiling error:
  CC [M]  drivers/media/i2c/pca9570.o
  In file included from drivers/media/i2c/pca9570.c:22:0:
  drivers/media/i2c/pca9570.c:339:25: error:
   'imx185_of_match' undeclared here (not in a function)
    MODULE_DEVICE_TABLE(of, imx185_of_match);

However, PCA9570 is an IO expender chip, which is
not related to imx185 and its of_match table.

So this patch just removes the imx185_of_match to
fix the compiling error.

Bug 200289382

Change-Id: I0e9aad2ea22b6592f320caeb7254ff42a3734a54
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1467896
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>