tegra: fd: Add wrapper for __alloc_fd
Modules like nvgpu and nvmap should to be able to allocate
all open fds and for this they need to call __alloc_fd()
rather than alloc_fd.
However, as __alloc_fd is not exported to modules, a wrapper
is being created and exported to the respective modules.

bug 200290850

Change-Id: Id245040842f71aadfc484711a23e8ce9dbfa872f
Signed-off-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1701552
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>