tegra: fd: Add wrapper for __alloc_fd
Modules like nvgpu and nvmap want to be able to allocate
all open fds and for this they need to call __alloc_fd()
rather than alloc_fd. However, as __alloc_fd is not
exported to modules, a wrapper is being created and
exported to the respective modules.

bug 200290850

Change-Id: I17dc0db90e1ebed0a264a83f6eb745ff756fb2dd
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/1468417
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>