t19x: dc: add function to toggle the SOR AFIFO
On T19x, an async FIFO is being added between the SOR
and the pad logic in order to meet certain timing
requirements given an 810MHz brick clock output (for HBR3).
During the SOR start sequence, the AFIFO must be unblocked
right before switching to the macro feedback clock. During
the SOR stop sequence, the AFIFO must be blocked right
before switching back to the safe clock. Unblocking
the AFIFO will allow data to flow between the SOR and pads,
and blocking the AFIFO will do the opposite.

This change adds a simple inline wrapper that writes to
the SOR_NV_PDISP_SOR_DP_LINKCTL0_0.ASYNC_FIFO_BLOCK field
in order to update the status of the AFIFO.

TDS-2129

Change-Id: Ia8703e845607febffa9688f592383603a5d85ead
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1463132
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aron Wong <awong@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>