ALSA: hda: build config(y/m) dependent code
when SND_HDA_TEGRA driver is build as module, the code depending
on this was not getting compiled, due to the conditional #if check
on the above config.

Modified the macro checks to complile when the config is either
defined as 'm' or 'y'.

Bug 200289390
Bug 200305194

Reviewed-on: http://git-master/r/1469651
(cherry picked from commit 91dbac0cc6519631ac73c18d19fb655bc6c1dd05)

Change-Id: Ie9942dd8cefe89eabdabd68127f10682cec27711
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: http://git-master/r/1476236
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>