pinctrl: t186: Use macros to define enums and pins
Currently, there is lots of duplication in the pin names for
defining unique ID, pincontrol descriptors, pins, function and
function name.

Make the table such that it can take the function argument and
define the unique ID, pincontrol descriptors and pins.

Similarly define the  table for function enum and name.

This will reduce lots of table/structure initialisation and reduce
the code side by 1600 lines.

Change-Id: I416370f625b0f28921cd1e7f088101ebdf80335d
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1470503
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>