video: tegra: hpd: fix DP hpd state machine
1. fix sending redundant hotplug events from kernel
to userspace in state machine by removing hpd_disable()
in hpd_plug_state()
- STATE_PLUG transition happens always from  STATE_HPD_RESET
and unplug event is already notified to userspace in
hpd_reset_state() in non-seamless scenario.
- STATE_PLUG transition happens from STATE_INIT_FROM_BOOTLOADER
and hwc is not required to be notified of unplug event if cable
is removed before state transition happens.
2. set switch state for hotplug after notifying hwc layer
of plug/unplug event.

Bug 200287104

Change-Id: I07440d336824b7083f63c9e5f8dfa1334bdd3cae
Signed-off-by: Santosh Reddy Galma <galmar@nvidia.com>
Reviewed-on: http://git-master/r/1470517
Reviewed-on: http://git-master/r/1474613
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>