platform: tegra: ari: resolve module dependence due to clear_serr symbol
When converting TEGRA_18X_SERROR to LKM, we found some symbols
dependence in many drivers. And which causes the modules can not be
inserted by changing any sequence.

This can be fixed by moving the "tegra18_clear_serr" symbol to ari_mca
driver and translating it to "ari_clear_serr". This is more likely the
right place that it should be.

Bug 200299501

Change-Id: I531e595decb1062f07d0f1cc6448c9eb089cf9f5
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1470915
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>