video: tegra: host: Ensure syncpt RAM init
On T194, syncpt RAM can be initialized by HW on poweron. It is necessary
for SW to ensure that syncpt RAM be used only after HW has finished
initializing it.

To that end, poll for SYNCPT_RAM_INIT_DONE on host1x poweron to which will
indicate that HW is done initializing syncpt RAM

Jira HOSTX-195

Change-Id: I1c64f9533005f0c9a2da0c07b77027b3b3214ccb
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1471012
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>