Asoc: tegra-alt: Support more sampling rates for Tegra186 M3420
The Tegra186 Orca-Viper (M3420) platform must support 44.1kHz, 48kHz
and 96kHz sampling rates. Update the machine driver for this platform
to support these rates. Support for 32kHz is also enabled for testing
although this is not strictly part of the requirement.

Please note that for Tegra186, in order to accurately configure the
audio PLL (pll_a) the PLL needs to be disabled, otherwise the p-divider
cannot be configured and this leads to inaccurate clock rates for some
sample rates. Therefore, enabling of the clocks is moved from the
machine start-up to when the 'hw_params' are configured and because of
this additional reference counting for enabling the clock is necessary.

Bug 200286098

Change-Id: I8a55b8fadcd08490c76d7ad40a1797508fd00a4e
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1472223
(cherry picked from commit 8cfc5d5d93eb7afa3fb85b9c1786e76cf59a26dd)
Reviewed-on: http://git-master/r/1477291
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>