gpu: nvgpu: Last fixes for kmem usage
Remove the last usage of vmalloc() that slipped through
the cracks. This was possible because depsite not including
the vmalloc.h header directly it was implicitly included.
Subsequent patches in this series fix the includes in the
kmem header files such that this type of mistake will now
be much harder.

Change-Id: Ibf20994097864283d4881340599ea446758cf5f2
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/1472369
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>