t194: display: SOR addrs for initial HDMI/DP PRODs
The initial HDMI/DP SOC PROD settings that we use for T194
will purely be derived from T186. However, the base offsets
for certain SOR registers are being shifted down on T194 due
to the addition of a new HEAD_STATE register. This patch updates
the offsets for those registers so that they're correct for
T194.

TDS-2131

Change-Id: I17eb5aabb4ab32774c757feb15865e295c15bcd5
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1473122
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Aron Wong <awong@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>