video: tegra: host: Set DMASTART and DMAEND properly
Currently host1x CDMA is programmed to allow it to
read from anywhere in the lower 32 bits of its
available memory space, even though its CDMA buffer
is only a small section of it.

Set the DMASTART and DMAEND pointers tightly to allow
CDMA only from the intended command buffer, and also
add support for 40-bit pointers on T186 while at it.

Bug 1919548

Change-Id: I78498fd703e0cc70f235531f962b8000a9fc4931
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1473588
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>