ASoC: tegra-alt: Fix I2S start-up for Orca-Viper
The Orca-Viper platform has 4 AD1937 codecs where each one is connected
to a different I2S channel. The platform has been designed so that the
4 I2S channel are synchronous with each other. This is achieved by
having I2S0 channel drive the bit-clock and frame-syncs for the other
3 I2S channels. Therefore, I2S0 must always be running to output audio
on any of the other channels.

One problem encountered with this, is that if one of the slave I2S
channels happens to start momentarily before I2S0, and so the bit-clock
is not yet active, the soft-reset of the I2S slave channel will fail
and cause audio playback on the slave channel to fail. To avoid this we
need to ensure that the bit-clock driven by I2S0 is active before we
start any slave channel. To way to accomplish this, is to detect which
I2S channel is the master when instantiating the card, add a PRE DAPM
widget that will run before playback starts to enable I2S master
bit-clock and a POST DAPM widget to disable the I2S master bit-clock
after playback stops.

Bug 200286098

Change-Id: I4dfba7720f271b0f36feafa84ff9276304db6198
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1475480
(cherry picked from commit ccec1edb4dc22c71e964174ffc55f17a680b545c)
Reviewed-on: http://git-master/r/1480806
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>