tegra: dc: Add support for dc client interface
For usecases where different kernel modules work
in tandem with dc driver and need to coordinate
with dc driver, define an interface where modules
dependent on dc driver will register as dc clients
and get notified for events they have registered
for. As a part of the registration process, the
client can register various callback functions per
head. dc driver allocates an unique client id per
head during registration.

In the first roolout, following are the supported
events for callback functions:
    - NOTIFY_DC_ENABLED_EVENT
    - NOTIFY_DC_DISABLED_EVENT
    - NOTIFY_MODESET_EVENT

However the interface is scalable and more events
can be added as and when needed.

Jira TDS-2026

Change-Id: I28c14572d445b9d65ef3ed1b7ca4e6b0e0bcf39f
Signed-off-by: Arun Swain <arswain@nvidia.com>
Reviewed-on: http://git-master/r/1474884
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>