ASoC: tegra-alt: Return error if DAI link setup fails
If the function that sets up the DAI links for the Orca-Viper platform
fails we do not return an error and continue on with the registration
of the card. There is no point registering the card if the DAI links
have not been successfully configured and so return an if setting up
the DAI links fail and allow the probe to fail gracefully.

Bug 200286098

Change-Id: I0b12746eeb04ec9319a72c0985e3f87f95ebad24
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1475479
(cherry picked from commit a704cc57b40c1d4747705b1c470b5622e45b7a7c)
Reviewed-on: http://git-master/r/1477288
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>