video: tegra: host: Clamp modules across reset
On T194, SW needs to clamp the signals between host1x and the client
module when the client goes through a reset sequence to prevent
meta-stability of the signals between host1x and the client.

Add some support that allows the clamp to be set/cleared across a module
reset.

Jira HOSTX-208

Change-Id: Ia58964d20ffd65299490f1dfbd788d661efa5180
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1476616
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>