trusty: fix race codition in trusty-ipc driver
There is a race codition in trusty-ipc driver that ipc_release()
could be called during the ipc_virtio_remove() when doing rmmod
trusty-virtio.

The ipc_release() function handles dn first and then chan while
the while loop in the _go_offline() handles chan first and then
dn. And there is not a comprehensive lock protection between two
of them. These result in a freed memory (dn) access when later
_go_offline() tries to shutdown this dn.

The ipc_release() is called via a system call from user space
who's accessing the cdev, which could happen in any time. So we
should make sure there is no ipc_release() being executed when
the driver tries to shutdown the dn in chan_trigger_event().

This patch sets ops_arg to NULL, prior to any other operation in
the ipc_release() so that ipc_virtio_remove() would notice that
this chan is in the middle of ipc_release() and wait for it.

Bug 200305192

Change-Id: I2863af89e43f37468866a67e3aed8e793fb6d6e5
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1480372
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>